完成lab1和lab2

+++ test/simpledb/HeapPageReadTest.java
@@ -96,8 +96,8 @@ public class HeapPageReadTest extends SimpleDbTestBase {
             IntField f0 = (IntField) tup.getField(0);
             IntField f1 = (IntField) tup.getField(1);
 
-            assertEquals(EXAMPLE_VALUES[row][0], f0.getValue());
-            assertEquals(EXAMPLE_VALUES[row][1], f1.getValue());
+            assertEquals((Integer) EXAMPLE_VALUES[row][0], f0.getValue());
+            assertEquals((Integer) EXAMPLE_VALUES[row][1], f1.getValue());
             row++;
         }
     }
