+++ test/simpledb/systemtest/AbortEvictionTest.java
@@ -34,7 +34,7 @@ public class AbortEvictionTest extends SimpleDbTestBase {
         insert.open();
         Tuple result = insert.next();
         assertEquals(SystemTestUtil.SINGLE_INT_DESCRIPTOR, result.getTupleDesc());
-        assertEquals(1, ((IntField)result.getField(0)).getValue());
+        assertEquals((Integer) 1, ((IntField)result.getField(0)).getValue());
         assertFalse(insert.hasNext());
         insert.close();
     }
