+++ test/simpledb/systemtest/EvictionTest.java
@@ -62,7 +62,7 @@ public class EvictionTest extends SimpleDbTestBase {
         insert.open();
         Tuple result = insert.next();
         assertEquals(SystemTestUtil.SINGLE_INT_DESCRIPTOR, result.getTupleDesc());
-        assertEquals(1, ((IntField)result.getField(0)).getValue());
+        assertEquals((Integer) 1, ((IntField)result.getField(0)).getValue());
         assertFalse(insert.hasNext());
         insert.close();
     }
